scraper/requirements.txt
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0
python-dotenv>=1.0.0
requests>=2.32.0
lxml>=5.0.0
rapidfuzz>=3.9.0
openai>=1.55.0
pydantic>=2.6.0
